made separate file for the function (#30)
